bookexchange_project/requirements.txt
django==4.1.4
djangorestframework==3.14.0
django-extra-fields==3.0.2
django-filter==22.1
djoser==2.1.0
gunicorn==20.1.0
pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
drf-yasg==1.21.5
